requirements.txt
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Faker==12.0.1
Flask==3.0.3
Flask-Admin==1.6.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mixer==7.2.2
peewee==3.17.3
pillow==10.3.0
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3
wtf-peewee==3.0.5
psycopg2-binary==2.9.9
